old-tests/submission/W3C/domhtml/HTMLScriptElement07.html

<!DOCTYPE html>
<html><head>
<meta charset="utf-8">
<title>script</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<link href="http://www.w3.org/TR/html5/obsolete.html#dom-script-event" rel="help">
<link rel="author" title="W3C">
</head>
<body>
<p>The event and htmlFor IDL attributes of the script element must return the empty string on getting, and do nothing on setting.</p>

<script>
var domtest = async_test("The event and htmlFor IDL attributes of the script element must return the empty string on getting, and do nothing on setting.");

on_event(window, "load", domtest.step_func(function (dom_event) {
    var nodeList;
      var testNode;
      var event;
      var doc;
      doc = window.document;
      nodeList = doc.getElementsByTagName("script");
      assert_equals(nodeList.length,3);
      testNode = nodeList.item(0);
      assert_equals(testNode.event, "");
      testNode.event = "dummy value";
      assert_equals(testNode.event, "");
      
    domtest.done();
}));

</script>
<div id="log"></div>
</body></html>
